+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/rpc/beaconchain/methods/BlobSidecarsByRootMessageHandlerTest.java
@@ -295,6 +295,6 @@ public class BlobSidecarsByRootMessageHandlerTest {
   }
 
   private int getMaxRequestBlobSidecars() {
-    return spec.getNetworkingConfigDeneb().orElseThrow().getMaxRequestBlobSidecars().intValue();
+    return spec.getNetworkingConfigDeneb().orElseThrow().getMaxRequestBlobSidecars();
   }
 }
